assets/requirements.txt
influxdb==5.2.2
jsonschema==3.0.2
paho-mqtt==1.4.0
pandas==0.25.0
requests==2.22.0
urllib3==1.25.3
qtconsole==4.5.2